Moved dtos to seperate folder
