+++ MyMoney.Web/FrontEnd/Angular/src/app/shared/api/http-helper.class.ts
@@ -14,7 +14,7 @@ export class HttpHelper extends CommonHttpHelper {
         super();
         
         this.sessionToken$ = store
-        .select(selectCurrentSession)
-        .pipe(map((session: ISessionModel | null) => session?.token ?? null));
+            .select(selectCurrentSession)
+            .pipe(map((session: ISessionModel | null) => session?.token ?? null));
     }
 }
