+++ MyMoney.Web/FrontEnd/Angular/src/app/shared/api/index.ts
@@ -3,31 +3,4 @@ export * from './budget.api';
 export * from './income.api';
 export * from './authentication.api';
 export * from './home.api';
-export {
-    IBudgetDto,
-    IIncomeDto,
-    IIncomeSearchDto,
-    IBudgetSearchDto,
-    IBudgetListDto,
-    IIncomeListDto,
-    IDateRangeDto,
-    IDeleteResultDto,
-    IIdDto,
-    ILoginDto,
-    ILoginResultDto,
-    IRegisterDto,
-    ITransactionDto,
-    ITransactionListDto,
-    IUpdateResultDto,
-    IRunningTotalSearchDto,
-    IRunningTotalDto,
-    IRunningTotalListDto,
-    IUserDto,
-    IBasicResultDto,
-    IForgotPasswordDto,
-    IPasswordDto,
-    Frequency,
-    IRecurringTransactionDto,
-    IRecurringEntityChildDto,
-    IRecurringIncomeDto,
-} from 'mymoney-common/lib/api';
+export * from 'mymoney-common/lib/api/dtos';
