+++ MyMoney.Web/FrontEnd/Common/src/api/dtos/dtos.interface.ts
