+++ MyMoney.Web/FrontEnd/Common/src/api/dtos/index.ts
@@ -1,0 +1,1 @@
+export * from './dtos.interface';
