+++ MyMoney.Web/FrontEnd/Common/src/api/index.ts
@@ -1,6 +1,5 @@
 export * from './transaction.api';
 export * from './budget.api';
-export * from './dtos.interface';
 export * from './income.api';
 export * from './authentication.api';
 export * from './home.api';
